Add: GA$

+++ database/migrations/2026_07_02_000001_create_order_notes_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         if (!Schema::hasTable('order_notes')) {
             Schema::create('order_notes', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedBigInteger('order_id');
+                $table->unsignedInteger('order_id');
                 $table->unsignedBigInteger('user_id')->nullable();
                 $table->text('content');
                 $table->string('type', 20)->default('info');
